1).gitignore(Убраны HEAD>>>) 2).travis почищен 3)Весь код в html.c, публичные функции в html.h 4) Переписаны GTest 5)В main.c ввод данных и печать структур
