+++ .travis.yml
@@ -33,7 +33,6 @@ before_script:
 script:
   - cmake ..
   - make || travis_terminate 1
-#  - valgrind --tool=memcheck --leak-check=full --show-leak-kinds=all --leak-resolution=med  --track-origins=yes ./test_main.out || travis_terminate 1
 
 after_success:
   - lcov --capture --directory . --output-file coverage.info
