+++ src/html.h
@@ -7,10 +7,11 @@
 typedef struct html_tag {
     char name[SIZE];
     char value[SIZE_VALUE];
+
 } html;
 
 int correct_name(const char * string, int count);
-html * html_decoder(const char * string);
-int start();
+html * html_decoder(const char * string, int * size);
 
 #endif //DZ1_MAIN_H
+
